% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dots.R
\name{add_dots}
\alias{add_dots}
\title{Add Dots (...) below kable Table}
\usage{
add_dots(kable_table)
}
\arguments{
\item{kable_table}{output from \code{kableExtra::kbl()}}
}
\value{
Object of class kableExtra
}
\description{
Add Dots (...) below kable Table
}
\examples{
iris \%>\%
 head() \%>\%
 kbl() \%>\%
 add_dots()
}
