% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rename_table.R
\name{assert_data_type}
\alias{assert_data_type}
\title{Assert data type}
\usage{
assert_data_type(x, any.missing = TRUE)
}
\arguments{
\item{x}{character vector to check}

\item{any.missing}{[\code{logical(1)}] \cr Are vectors with missing values allowed? Default is TRUE.}
}
\description{
data_type is a special column in the data archive pipeline. This function
allows to check if it contains only allowed abbreviation (and optional no NA)
}
