% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_rename_fun.R
\name{check_data_pipeline_arguments}
\alias{check_data_pipeline_arguments}
\title{Check arguments of the data pipeline: l, convention_table and column_headers}
\usage{
check_data_pipeline_arguments(
  l,
  convention_table,
  column_headers,
  check_l_ch_consistency = TRUE
)
}
\arguments{
\item{l}{list containing all tibbles}

\item{convention_table}{tibble with conventional variables}

\item{column_headers}{tibble with column header descriptions}

\item{check_l_ch_consistency}{[\code{logical(1)}] \cr If consistency between l and column_headers should be checked}
}
\description{
Check arguments of the data pipeline: l, convention_table and column_headers
}
