% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_rename_fun.R
\name{column_rename_fun}
\alias{column_rename_fun}
\title{Column Rename Function}
\usage{
column_rename_fun(
  l,
  column_headers_path,
  convention_table_path = NULL,
  verbose_factor_levels = TRUE
)
}
\arguments{
\item{l}{list containing tibbles}

\item{column_headers_path}{path to file created by \code{\link{create_rename_table}}}

\item{convention_table_path}{path to file containing conventions (e.g. \code{C:/Users/858782/DSM/BovaerAnimalTrials - General/01_Admin/Animal_Trial_Column_Names_RZ.xlsx})}

\item{verbose_factor_levels}{Boolian, should automatically determined factor levels be printed to console?}
}
\description{
Column Rename Function
}
