% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_excel_archive.R
\name{create_excel_archive}
\alias{create_excel_archive}
\title{Create Excel Archive}
\usage{
create_excel_archive(
  l,
  filename = "data_archive",
  convention_table_path = NULL,
  column_headers_path = NULL,
  ERD = TRUE,
  pk = NULL,
  connectors = NULL,
  erd_height = 800
)
}
\arguments{
\item{l}{list containing tibbles}

\item{filename}{path and filename (without suffix) of the file to create}

\item{convention_table_path}{path to file containing conventions (e.g. \code{C:/Users/858782/DSM/BovaerAnimalTrials - General/01_Admin/Animal_Trial_Column_Names_RZ.xlsx})}

\item{column_headers_path}{path to file created by \code{\link{create_rename_table}}}

\item{ERD}{Boolean, should Entity-Relationship-Diagram be drawn and added as tab?}

\item{pk}{list defining the primary keys per table as character string}

\item{connectors}{character vector defining the connections between tables}

\item{erd_height}{height of the ERD plot in pixels (manually adjust if necessary)}
}
\description{
Create Excel Archive
}
