% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rename_table.R
\name{create_rename_table}
\alias{create_rename_table}
\title{Create Rename Table}
\usage{
create_rename_table(l, file = "column_headers.xlsx", overwrite = FALSE)
}
\arguments{
\item{l}{list containing tibbles}

\item{file}{path, filename and suffix of the file to create}

\item{overwrite}{Boolian, should an existing file be overwritten?}
}
\description{
Create Rename Table
}
