% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunAxisPOSIXct.R
\name{FunAxisPOSIXct}
\alias{FunAxisPOSIXct}
\title{Axis Function for POSIXct}
\usage{
FunAxisPOSIXct(
  xtitle = "x axis",
  ytitle = "y axis",
  xline = 2,
  yline = 2,
  xpadj = 0,
  ypadj = 0,
  xcextitle = par("cex"),
  ycextitle = par("cex"),
  xcexlab = par("cex"),
  ycexlab = par("cex"),
  xat = NULL,
  yat = NULL,
  xlabels = TRUE,
  ylabels = TRUE,
  xtick = TRUE,
  ytick = TRUE,
  xtick_lwd = 1.5,
  ytick_lwd = 1.5,
  x_mtext = list(),
  y_mtext = list(),
  x_axis_1 = list(),
  y_axis_1 = list(),
  x_axis_2 = list(),
  y_axis_2 = list(),
  xmt = T,
  ymt = T,
  xmtn = 2,
  ymtn = 2,
  xmtratio = 0.5,
  ymtratio = 0.5,
  xmtbothsides = T,
  ymtbothsides = T,
  xmt_args = list(),
  ymt_args = list(),
  box_lwd = 1.5,
  box_args = list(),
  xtimevec,
  format = ""
)
}
\arguments{
\item{xtitle}{x axis title as string}

\item{xline}{space between x axis title and plotting region in mex}

\item{xpadj}{space between x axis title and plotting region may be
disturbed if you use bquote(). Adapt space with e.g. padj = 0.1,
padj = 0 means right or top alignment, and padj = 1 means left or bottom alignment.}

\item{xcextitle}{x axis title character extension (font size)}

\item{xcexlab}{x axis labels character extension (font size)}

\item{xat}{the points at which x axis tick-marks are to be drawn.}

\item{xlabels}{logical or character vector of labels to be placed at the tickpoints.}

\item{xtick}{a logical value specifying whether tickmarks and an axis line should
be drawn.}

\item{xtick_lwd}{line width of ticks on the x axis}

\item{x_mtext}{list of arguments passed to bottom x axis' mtext()
e.g. outer, adj, col, font, ...}

\item{x_axis_1}{list of arguments passed to bottom x axis' axis()
e.g. las, pos, outer, font, lty, lwd, lwd.ticks, col, col.ticks, hadj, ...}

\item{x_axis_2}{list of arguments passed to top x axis' axis()}

\item{xmt}{logical, should minor ticks on x axis be plotted}

\item{xmtn}{number of intervals in which to divide the area between major tick marks on the X-axis.
Set to 1 to suppress minor tick marks.}

\item{xmtratio}{ratio of lengths of minor x tick marks to major x tick marks.
The length of major tick marks is retrieved from par("tck").}

\item{xmtbothsides}{should minor ticks be plotted on both x axis (bottom and top)}

\item{xmt_args}{list of arguments passed to x axis' axis() to plot minor ticks.
e.g. lwd, lty, col, ...}

\item{box_lwd}{line width of box arround the plot}

\item{box_args}{list of arguments passed to box() to plot box arround plot.
e.g. which, lty, ...}

\item{xtimevec}{vector with all date-times to produce the plot (x argument of the plot function)}

\item{format}{character vector specifying the way to print the date-time e.g. "\\%m.\\%y". See strptime.}

\item{y...}{each argument for the y axis}
}
\value{
axis, ticks, labels and axistitle
}
\description{
Make axis, ticks, lables and axistitle in one step for timeseries.
Attention: axis.POSIXct is currently only implemented for the x axis.
}
\examples{
library(lubridate)
x <- now() + days(0:7)
y <- rnorm(8)

plot(x,y, type = "l", axes = F, ann = F)
FunAxisPOSIXct(xtimevec = x, format = "\%d.\%m")
}
\seealso{
box, mtext, axis
}
\author{
Reto Zihlmann
}
