% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunFactorize.R
\name{FunFactorize}
\alias{FunFactorize}
\title{Factorize columns}
\usage{
FunFactorize(DataSetName, VariableNames)
}
\arguments{
\item{DataSetNames}{Name of the dataframe where some columns should be factorized}

\item{Variablenames}{string vector, Name of the columns to be factorized}
}
\value{
Dataframe with factorized columns
}
\description{
Factorize columns which were not identified as factors
}
\examples{
df <- data.frame(x = 1:10, y = letters[1:10])
str(df)
df$y <- as.character(df$y)
str(df)
str(FunFactorize(df, "y"))
}
