% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunPar.R
\name{FunPar}
\alias{FunPar}
\title{Preferred par values funktion}
\usage{
FunPar(mar = c(3.5, 3.5, 2, 2), xaxs = "i", yaxs = "i", tcl = 0.4,
  las = 1, mgp = c(2, 0.5, 0), lab = c(4, 4, 7), family = "Segoe",
  cex = 0.9)
}
\arguments{
\item{mar}{smaller plot margin}

\item{xaxs}{do not add 4 percent to x axis' lim}

\item{yaxs}{do not add 4 percent to y axis' lim}

\item{tcl}{ticks look inside}

\item{las}{turn y axis labels}

\item{mgp}{position of axis titles and labels in lines}

\item{lab}{preferd number of ticks (4 instead of 5)}
}
\value{
invisible original par values
}
\description{
Set parameters (par) in R to favorit default values. Returns previous par values
invisible.
}
\examples{
plot(1:10)           # plotting with original par settings
opar <- FunPar()     # set new par settings
plot(1:10)
par(opar)            # reset original par settings
plot(1:10)
}
\seealso{
par()
}
\author{
Reto Zihlmann
}
