% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunPlotPosition.R
\name{FunPlotPosition}
\alias{FunPlotPosition}
\title{Get absolute plot coordinates from relative plot coordinates}
\usage{
FunPlotPosition(edge = c(0.1, 0.9))
}
\arguments{
\item{edge}{vector c(x,y) with the relative coordinate given in the range of [0,1]}
}
\value{
vector c(x,y) with the absolute coordinates for the current plot
}
\description{
The function transformes relative plot coordinates to absolute coordinates for the current plot.
}
\examples{
plot(1:10, 21:30)
text(FunPlotPosition()[1], FunPlotPosition()[2], "Hello")
FunPlotPosition()
}
