% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunRZ-package.R
\docType{package}
\name{FunRZ-package}
\alias{FunRZ}
\alias{FunRZ-package}
\title{FunRZ: Useful functions}
\description{
This package provides a wild collection of functions for data processing and visualization.
}
\keyword{internal}
