% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunSigStar.R
\name{FunSigStar}
\alias{FunSigStar}
\title{Significance Star Function}
\usage{
FunSigStar(pval)
}
\arguments{
\item{pval}{A p value to be expressed in asterisk notation}
}
\value{
asterisk
}
\description{
Takes any vector of p values and returnes the vector in asterisk notation
with p<0.05 (\emph{), <0.01 (\strong{), and <0.001 (}})
}
\examples{
FunSigStar(0.01)
FunSigStar(c(0.02, 0.05, 0.06))
}
