% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunTukeyHSD.R
\name{FunTukeyHSD}
\alias{FunTukeyHSD}
\title{boxplots with cld}
\usage{
FunTukeyHSD(
  mod = y ~ x,
  data = dat,
  upp_mar = 4,
  col = NA,
  ylab = "",
  xlab = "",
  main = "",
  las = 3,
  add_legend = F,
  legend_pos = "topright",
  legend = NULL,
  fill = "white"
)
}
\arguments{
\item{mod}{model passed to aov() with value ~ factor}

\item{data}{data.frame with the variables}

\item{upp_mar}{specify the number of lines on the top of the plot.
If there are many letters extra space is necessary}

\item{col}{vector with color to fill the boxes of the boxplot}

\item{ylab}{y axis title}

\item{xlab}{x axis title}

\item{main}{main title}

\item{add_legend}{logical if a legend should be add}

\item{legend_pos}{position specified as character (right, topleft,...)}

\item{legend}{string vector of length ≥ 1 to appear in the legend}

\item{fill}{vector with color to fill boxes in legend}

\item{orientation}{of axis labels see par()}
}
\value{
boxplot with cld
}
\description{
plot boxplots with compact letter display of pairwise comparisons
}
\examples{
df <- data.frame(x = c(rnorm(80, 10,2), rnorm(20, 7, 2)),
myfac = c(rep(c("a","b", "c", "d"), 20), rep("e", 20)))
FunTukeyHSD(mod = x ~ myfac, data = df, upp_mar = 5, col = c(rep("red", 3), "blue", "blue"),
            ylab = "y axis", las = 2, add_legend = T, legend = c("treatment", "control"),
            fill = c("blue", "red"))
}
