% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunXPosition.R
\name{FunXPosition}
\alias{FunXPosition}
\title{Get absolute x coordinates from relative x coordinates}
\usage{
FunXPosition(rel = 0.1)
}
\arguments{
\item{rel}{relative x coordinate given in the range of [0,1]}
}
\value{
absolute x coordinate for the current plot
}
\description{
The function transformes relative x coordinates to absolute x coordinates for the current plot.
}
\examples{
plot(1:10, 21:30)
FunXPosition(0.5)
}
