% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunYPosition.R
\name{FunYPosition}
\alias{FunYPosition}
\title{Get absolute y coordinates from relative y coordinates}
\usage{
FunYPosition(rel = 0.1)
}
\arguments{
\item{rel}{relative y coordinate given in the range of [0,1]}
}
\value{
absolute y coordinate for the current plot
}
\description{
The function transformes relative y coordinates to absolute y coordinates for the current plot.
}
\examples{
plot(1:10, 21:30)
FunYPosition(0.5)
}
