% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_theme_sci.R
\name{gt_theme_sci}
\alias{gt_theme_sci}
\title{Add scientific theme to gt}
\usage{
gt_theme_sci(data)
}
\arguments{
\item{dat}{The gt table data object. This is the gt table object that is commonly created through use of the gt() function.}
}
\value{
object of class \code{gt_tbl}
}
\description{
Add scientific theme to gt
}
\examples{
iris \%>\%
 head() \%>\%
 gt() \%>\%
 gt_theme_sci()
}
