% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_corresponding_html.R
\name{open_corresponding_html}
\alias{open_corresponding_html}
\title{Open corresponding html of a Rmd file}
\usage{
open_corresponding_html()
}
\description{
Opens corresponding html file of an open Rmd file in Viewer pane
without recompiling it
}
