% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_cor.R
\name{panel.cor}
\alias{panel.cor}
\title{plot correlation}
\usage{
panel.cor(x, y, digits = 2, prefix = "", cex.cor, text_pos = c(0.5, 0.5), ...)
}
\arguments{
\item{x, y}{Variables to calculate correlation}

\item{digits}{integer indicating the number of decimal places to be used}

\item{prefix}{character what should be put before number}

\item{cex.cor}{character expansion of the plotted correlation}

\item{tex_pos}{vector in the form c(x,y), defines the position of
the correlation. The coordinates are given as
normalized device coordinates.}
}
\value{
plot correlation
}
\description{
panel function to plot correlation in pair plot
}
\examples{
pairs(iris, lower.panel = panel.smooth, diag.panel = panel.hist,
      upper.panel = panel.cor)
}
