% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_hist.R
\name{panel.hist}
\alias{panel.hist}
\title{plot histograms in pair plot}
\usage{
panel.hist(x, ...)
}
\arguments{
\item{x}{Variable for which a histogram should be plotted}
}
\value{
plot a histogram
}
\description{
panel function to plot histograms in pair plot
}
\examples{
pairs(iris, lower.panel = panel.smooth, diag.panel = panel.hist,
      upper.panel = panel.cor)
}
