% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj_temp.R
\name{proj_temp}
\alias{proj_temp}
\title{Create Project Folder Structure}
\usage{
proj_temp()
}
\value{
Returns invisible logical vector indicating if folder creation was successful
}
\description{
Create template folder structure for RStudio project. Existing folders are not overwritten.
}
