% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_objects.R
\docType{data}
\name{RCol}
\alias{RCol}
\title{Reto Colors (v3)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25 rows and 6 columns.
}
\usage{
RCol
}
\description{
Reto colors grouped by color (gree, orange, yellow, red, blue) and hue (1-5)
}
\examples{
png("RCol_v3.png", width = 11.4, height = 11.4, units = "cm", res = 300)
par(mar = c(5,5,1,1))
with(RCol, plot(x = as.numeric(factor(color)), y = hue, bg = hex,
                cex = 9, pch = 22, xlim = c(0.6,5.4),
                ylim = c(0.6,5.4), xaxt = "n", xlab = "color"))
with(RCol, text(x = as.numeric(factor(color)),
                y = hue-0.3, labels = hex, cex = 0.3))
axis(1, at = 1:5, labels = c("green", "orange", "yellow", "red", "blue"))
dev.off()
}
\keyword{datasets}
