% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart_r_session.R
\name{restart_r_session}
\alias{restart_r_session}
\title{Restart R session}
\usage{
restart_r_session()
}
\description{
Like the Ctrl+Shift+F10 command but additionally clears console and
moves focus to source editor
}
