% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rename_table.R
\name{style_data_type}
\alias{style_data_type}
\title{Style Excel Cell containing Data Type}
\usage{
style_data_type(wb, sheet, cols, rows)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{Columns to apply conditional formatting to}

\item{rows}{Rows to apply conditional formatting to}
}
\description{
Style Excel Cell containing Data Type
}
