% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rename_table.R
\name{style_logical}
\alias{style_logical}
\title{Style Excel Cells TRUE/FALSE}
\usage{
style_logical(wb, sheet, cols, rows)
}
\arguments{
\item{wb}{A workbook object}

\item{sheet}{A name or index of a worksheet}

\item{cols}{Columns to apply conditional formatting to}

\item{rows}{Rows to apply conditional formatting to}
}
\description{
Style Excel Cells TRUE/FALSE
}
