% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_add_sensitivity.R
\name{wb_add_sensitivity}
\alias{wb_add_sensitivity}
\title{Add Companty Sensitivity Label}
\usage{
wb_add_sensitivity(
  wb,
  sensitivity = c("internal", "public", "confidential", "secret")
)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object}

\item{sensitivity}{Sensitivity label, one of the following: "public", "internal" (default), "confidential", "secret"}
}
\value{
A \code{wbWorkbook} object
}
\description{
Add Companty Sensitivity Label
}
\examples{
wb \%>\% wb_add_sensitivity(sensitivity = "secret")
}
