% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_format_header.R
\name{wb_format_header}
\alias{wb_format_header}
\title{Format Header of Excel Table}
\usage{
wb_format_header(wb, sheet = openxlsx2::current_sheet(), dat)
}
\arguments{
\item{wb}{A \code{wbWorkbook} object}

\item{sheet}{A name or index of a worksheet}

\item{dat}{data which were included in Excel sheet, used to extract dimensions of header}
}
\value{
A \code{wbWorkbook} object
}
\description{
Format Header of Excel Table
}
\examples{
wb_workbook(creator = "Reto Zihlmann") \%>\%
  wb_add_worksheet(sheet = "README", tab_color = "green") \%>\%
  wb_add_data(sheet = 1, x = dat, with_filter = FALSE) \%>\%
  wb_set_col_widths(sheet = 1, cols = 1:ncol(dat), widths = "auto") \%>\%
  wb_format_header(dat = dat)
}
