% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rename_table.R
\name{write_column_headers_sheet}
\alias{write_column_headers_sheet}
\title{Write sheet containing column headers}
\usage{
write_column_headers_sheet(
  column_headers,
  tabname = "column_headers",
  for_rename_table = FALSE
)
}
\arguments{
\item{column_headers}{column_headers tibble}

\item{tabname}{Name of the Tab which is added to xlsx sheet}

\item{for_rename_table}{Boolean, is the tab for a rename table (containing has_convention, keep, ...) or not}
}
\description{
Write sheet containing column headers
}
