% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xlsx.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Write data to an xlsx file}
\usage{
write_xlsx(
  x,
  file,
  as_table = FALSE,
  col_widths = "auto",
  sensitivity = "internal",
  ...
)
}
\arguments{
\item{x}{see \code{\link[openxlsx2:write_xlsx]{openxlsx2::write_xlsx()}}}

\item{file}{see \code{\link[openxlsx2:write_xlsx]{openxlsx2::write_xlsx()}}}

\item{as_table}{see \code{\link[openxlsx2:write_xlsx]{openxlsx2::write_xlsx()}}}

\item{col_widths}{see \code{\link[openxlsx2:write_xlsx]{openxlsx2::write_xlsx()}}}

\item{sensitivity}{sensitivity Sensitivity label, one of the following: "public", "internal" (default), "confidential", "secret"}

\item{...}{see \code{\link[openxlsx2:write_xlsx]{openxlsx2::write_xlsx()}}}
}
\value{
the \code{wbWorkbook} object, invisibly
}
\description{
Write a data frame or list of data frames to an xlsx file.
}
\examples{
write_xlsx(x = list(iris = iris, mtcars = mtcars), file = "~/Downloads/test_to_delete_write.xlsx")
}
