% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dots.R
\name{add_dots}
\alias{add_dots}
\title{Add Dots (...) below display Table}
\usage{
add_dots(input_table)
}
\arguments{
\item{input_table}{output from \code{kableExtra::kbl()} or \code{gt::gt()} or \code{knitr::kable()}}
}
\value{
Object of class kableExtra or output is directly printed as html?
}
\description{
This function has become a bit too complex, maybe it's easier to just say 'we print the first few rows of the dataset'
}
\examples{
iris \%>\%
 head() \%>\%
 kableExtra::kbl() \%>\%
 add_dots()
}
