% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunPar.R
\name{FunPar}
\alias{FunPar}
\title{Preferred par values funktion}
\usage{
FunPar(
  mar = c(3.5, 3.5, 2, 2),
  xaxs = "i",
  yaxs = "i",
  tcl = 0.4,
  las = 1,
  mgp = c(2, 0.5, 0),
  lab = c(4, 4, 7),
  family = "Segoe",
  cex = 0.9
)
}
\arguments{
\item{mar}{smaller plot margin. default c(5.1, 4.1, 4.1, 2.1))}

\item{xaxs}{do not add 4 percent to x axis' lim. default "r"}

\item{yaxs}{do not add 4 percent to y axis' lim. default "r"}

\item{tcl}{ticks look inside. default -0.5}

\item{las}{turn y axis labels. default 0}

\item{mgp}{position of axis titles and labels in lines. default c(3, 1, 0)}

\item{lab}{preferd number of ticks (4 instead of 5). default c(5, 5, 7)}

\item{family}{prefered font family (Segoe). default ""}

\item{cex}{prefered character expansion factor. default 1}
}
\value{
invisible original par values
}
\description{
Set parameters (par) in R to favorit default values. Returns previous par values
invisible.
}
\examples{
plot(1:10)           # plotting with original par settings
opar <- FunPar()     # set new par settings
plot(1:10)
par(opar)            # reset original par settings
plot(1:10)
}
\seealso{
par()
}
\author{
Reto Zihlmann
}
